refactor: moved glide service location -> renamed to glideutils
