--- data/src/main/java/com/haomins/data/service/ReaderGlideModule.kt
@@ -1,14 +1,0 @@
-package com.haomins.data.service
-
-import android.content.Context
-import android.util.Log
-import com.bumptech.glide.GlideBuilder
-import com.bumptech.glide.annotation.GlideModule
-import com.bumptech.glide.module.AppGlideModule
-
-@GlideModule
-class ReaderGlideModule : AppGlideModule() {
-    override fun applyOptions(context: Context, builder: GlideBuilder) {
-        builder.setLogLevel(Log.ERROR)
-    }
-}
